Front-end overhaul to optimise mobile view

+++ be/Services/TorrentService.cs
@@ -30,15 +30,15 @@ public class TorrentService(
 
     private ClientEngine? engine;
 
-    private readonly Dictionary<string, TorrentManager> torrents = [];
+    private readonly List<(string InfoHash, TorrentManager Manager)> managers = [];
 
     /// <summary>
     /// Get all saved torrents
     /// </summary>
     /// <returns>The saved torrents</returns>
     public IEnumerable<Dtos.Torrent> GetTorrents()
-        => torrents
-            .Select(kv => Dtos.Torrent.ConvertFrom(kv.Key, kv.Value))
+        => managers
+            .Select(pair => Dtos.Torrent.ConvertFrom(pair.InfoHash, pair.Manager))
             .ToArray();
 
     /// <summary>
@@ -48,13 +48,14 @@ public class TorrentService(
     /// <returns>The saved torrent</returns>
     public Dtos.Torrent? GetTorrent(string infoHash)
     {
-        if (!torrents.TryGetValue(infoHash, out var torrent))
+        var manager = managers.SingleOrDefault(pair => pair.InfoHash == infoHash).Manager;
+        if (manager is null)
         {
             logger.LogWarning("Torrent {infoHash} not found", infoHash);
             return null;
         }
 
-        var response = Dtos.Torrent.ConvertFrom(infoHash, torrent);
+        var response = Dtos.Torrent.ConvertFrom(infoHash, manager);
         return response;
     }
 
@@ -65,7 +66,7 @@ public class TorrentService(
     /// <returns>Whether the torrent could be saved</returns>
     public bool SaveTorrent(string infoHash)
     {
-        if (torrents.ContainsKey(infoHash))
+        if (managers.Any(manager => manager.InfoHash == infoHash))
         {
             logger.LogWarning("Torrent {infoHash} already saved", infoHash);
             return false;
@@ -83,13 +84,14 @@ public class TorrentService(
     /// <returns>Whether the request could be accepted</returns>
     public bool RemoveTorrent(string infoHash)
     {
-        if (!torrents.TryGetValue(infoHash, out var torrent))
+        var manager = managers.SingleOrDefault(pair => pair.InfoHash == infoHash).Manager;
+        if (manager is null)
         {
             logger.LogWarning("Torrent {infoHash} not found", infoHash);
             return false;
         }
 
-        Task.Run(() => DeleteTorrentAsync(infoHash, torrent));
+        Task.Run(() => DeleteTorrentAsync(infoHash, manager));
 
         return true;
     }
@@ -102,13 +104,14 @@ public class TorrentService(
     /// <returns>Whether the request could be accepted</returns>
     public bool UpdateTorrent(string infoHash, TorrentPatch patch)
     {
-        if (!torrents.TryGetValue(infoHash, out var torrent))
+        var manager = managers.SingleOrDefault(pair => pair.InfoHash == infoHash).Manager;
+        if (manager is null)
         {
             logger.LogWarning("Torrent {infoHash} not found", infoHash);
             return false;
         }
 
-        Task.Run(() => UpdateTorrentAsync(infoHash, torrent, patch));
+        Task.Run(() => UpdateTorrentAsync(infoHash, manager, patch));
 
         return true;
     }
@@ -136,38 +139,36 @@ public class TorrentService(
             }
             catch (Exception ex)
             {
-                logger.LogError(ex, "An error occurred loading torrent {file}: {reason}", file, ex.Message);
+                logger.LogError(ex, "An error occurred loading torrent {file}: {reason}", file, ex);
             }
         }));
 
         // Monitor torrents
         while (!stoppingToken.IsCancellationRequested)
         {
-            var torrentsArray = torrents.ToArray();
-            foreach (var kv in torrentsArray)
+            var managersArray = managers.ToArray();
+            foreach (var (infoHash, manager) in managersArray)
             {
                 try
                 {
-                    var infoHash = kv.Key;
-                    var torrent = kv.Value;
-                    var progress = torrent.PartialProgress;
-                    switch (torrent.State)
+                    var progress = manager.PartialProgress;
+                    switch (manager.State)
                     {
                         case TorrentState.Seeding:
                             progress = 100;
                             break;
                         case TorrentState.Error:
-                            await torrent.StopAsync();
-                            await torrent.StartAsync();
+                            await manager.StopAsync();
+                            await manager.StartAsync();
                             break;
                         case TorrentState.Stopped:
                             continue;
                     }
-                    if (progress == 100) await CompleteTorrentAsync(infoHash, torrent);
+                    if (progress == 100) await CompleteTorrentAsync(infoHash, manager);
                 }
                 catch (Exception ex)
                 {
-                    logger.LogError(ex, "An error occurred while monitoring torrent {infoHash}: {reason}", kv.Key, ex.Message);
+                    logger.LogError(ex, "An error occurred while monitoring torrent {infoHash}: {reason}", infoHash, ex);
                 }
             }
             await Task.Delay(1_000, stoppingToken);
@@ -178,111 +179,112 @@ public class TorrentService(
     {
         try
         {
-            if (torrents.TryGetValue(infoHash, out var torrent)) return torrent;
+            var manager = managers.SingleOrDefault(pair => pair.InfoHash == infoHash).Manager;
+            if (manager is not null) return manager;
 
             var magnetLink = new MagnetLink(InfoHash.FromHex(infoHash));
-            var manager = await engine!.AddAsync(magnetLink, InProgressDirectory);
+            manager = await engine!.AddAsync(magnetLink, InProgressDirectory);
             await manager.StartAsync();
-            torrents[infoHash] = manager;
+            managers.Add((infoHash, manager));
             logger.LogInformation("Downloading {infoHash} to \"{path}\"", infoHash, InProgressDirectory);
             return manager;
         }
         catch (Exception ex)
         {
-            logger.LogError(ex, "An error occurred downloading {infoHash}: {reason}", infoHash, ex.Message);
+            logger.LogError(ex, "An error occurred downloading {infoHash}: {reason}", infoHash, ex);
             throw;
         }
     }
 
-    private async Task DeleteTorrentAsync(string infoHash, TorrentManager torrent)
+    private async Task DeleteTorrentAsync(string infoHash, TorrentManager manager)
     {
         try
         {
-            if (torrent.State != TorrentState.Stopped)
+            if (manager.State != TorrentState.Stopped)
             {
-                await torrent.StopAsync();
+                await manager.StopAsync();
 
-                foreach (var file in torrent.Files)
+                foreach (var file in manager.Files)
                 {
-                    File.Delete(file.FullPath);
+                    if (File.Exists(file.FullPath)) File.Delete(file.FullPath);
                     logger.LogDebug("Deleted \"{path}\"", file.FullPath);
                 }
 
-                DeleteDirectories(torrent);
+                DeleteDirectories(manager);
 
                 var torrentFile = Path.Join(MetadataDirectory, $"{infoHash}.torrent");
-                File.Delete(torrentFile);
+                if (File.Exists(torrentFile)) File.Delete(torrentFile);
                 logger.LogDebug("Deleted \"{path}\"", torrentFile);
 
-                await engine!.RemoveAsync(torrent);
+                await engine!.RemoveAsync(manager);
             }
 
-            torrents.Remove(infoHash);
+            managers.Remove((infoHash, manager));
 
-            logger.LogInformation("Removed torrent \"{name} ({infoHash})", torrent.Name, infoHash);
+            logger.LogInformation("Removed torrent \"{name} ({infoHash})", manager.Name, infoHash);
         }
         catch (Exception ex)
         {
-            logger.LogError(ex, "An error occurred removing {infoHash}: {reason}", infoHash, ex.Message);
+            logger.LogError(ex, "An error occurred removing {infoHash}: {reason}", infoHash, ex);
             throw;
         }
     }
 
-    private async Task UpdateTorrentAsync(string infoHash, TorrentManager torrent, TorrentPatch patch)
+    private async Task UpdateTorrentAsync(string infoHash, TorrentManager manager, TorrentPatch patch)
     {
         try
         {
             switch (patch.State)
             {
                 case PatchState.Downloading:
-                    await torrent.StartAsync();
-                    logger.LogInformation("Resuming torrent \"{name}\" ({infoHash})", torrent.Name, infoHash);
+                    await manager.StartAsync();
+                    logger.LogInformation("Resuming torrent \"{name}\" ({infoHash})", manager.Name, infoHash);
                     break;
                 case PatchState.Paused:
-                    await torrent.PauseAsync();
-                    logger.LogInformation("Pausing torrent \"{name}\" ({infoHash})", torrent.Name, infoHash);
+                    await manager.PauseAsync();
+                    logger.LogInformation("Pausing torrent \"{name}\" ({infoHash})", manager.Name, infoHash);
                     break;
             }
 
             foreach (var file in patch.Files ?? [])
             {
-                await torrent.SetFilePriorityAsync(
-                    torrent.Files.Single(torrentFile => torrentFile.Path == file.Path),
+                await manager.SetFilePriorityAsync(
+                    manager.Files.Single(torrentFile => torrentFile.Path == file.Path),
                     Convert(file.Priority));
-                logger.LogInformation("Setting priority to {priority} for \"{path}\" of torrent \"{name}\" ({infoHash})", file.Priority, file.Path, torrent.Name, infoHash);
+                logger.LogInformation("Setting priority to {priority} for \"{path}\" of torrent \"{name}\" ({infoHash})", file.Priority, file.Path, manager.Name, infoHash);
             }
         }
         catch (Exception ex)
         {
-            logger.LogError(ex, "An error occurred updating {infoHash}: {reason}", infoHash, ex.Message);
+            logger.LogError(ex, "An error occurred updating {infoHash}: {reason}", infoHash, ex);
             throw;
         }
     }
 
-    private async Task CompleteTorrentAsync(string infoHash, TorrentManager torrent)
+    private async Task CompleteTorrentAsync(string infoHash, TorrentManager manager)
     {
         try
         {
-            var newPath = Path.Join(CompletedDirectory, torrent.Name);
-            await torrent.StopAsync();
-            await torrent.MoveFilesAsync(newPath, true);
+            var newPath = Path.Join(CompletedDirectory, manager.Name);
+            await manager.StopAsync();
+            await manager.MoveFilesAsync(newPath, true);
             logger.LogDebug("Moved files to \"{path}\"", newPath);
-            DeleteUnwantedFiles(torrent);
-            DeleteDirectories(torrent);
-            await engine!.RemoveAsync(torrent);
+            DeleteUnwantedFiles(manager);
+            DeleteDirectories(manager);
+            await engine!.RemoveAsync(manager);
 
-            logger.LogInformation("Completed torrent \"{name} ({infoHash})", torrent.Name, infoHash);
+            logger.LogInformation("Completed torrent \"{name} ({infoHash})", manager.Name, infoHash);
         }
         catch (Exception ex)
         {
-            logger.LogError(ex, "An error occurred completing {infoHash}: {reason}", infoHash, ex.Message);
+            logger.LogError(ex, "An error occurred completing {infoHash}: {reason}", infoHash, ex);
             throw;
         }
     }
 
-    private void DeleteUnwantedFiles(TorrentManager torrent)
+    private void DeleteUnwantedFiles(TorrentManager manager)
     {
-        foreach (var file in torrent.Files)
+        foreach (var file in manager.Files)
         {
             if (file.Priority == MonoTorrent.Priority.DoNotDownload)
             {
@@ -292,9 +294,9 @@ public class TorrentService(
         }
     }
 
-    private void DeleteDirectories(TorrentManager torrent)
+    private void DeleteDirectories(TorrentManager manager)
     {
-        var directory = Path.Join(InProgressDirectory, torrent.Name);
+        var directory = Path.Join(InProgressDirectory, manager.Name);
         if (Directory.Exists(directory)) Directory.Delete(directory, true);
     }
